Capstone/pyproject.toml
[tool.poetry]
name = "capstone"
version = "0.1.0"
description = ""
authors = ["Derek <derekb0147@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.5"
requests = "^2.23.0"
django-crispy-forms = "^1.9.0"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
